Move API clients/adapters from `kopf.k8s` to `kopf.clients`
