+++ kopf/clients/fetching.py
