+++ kopf/clients/patching.py
