+++ tests/k8s/test_watching.py
@@ -12,7 +12,7 @@ import logging
 
 import pytest
 
-from kopf.k8s.watching import streaming_watch, infinite_watch, WatchingError
+from kopf.clients.watching import streaming_watch, infinite_watch, WatchingError
 
 STREAM_WITH_NORMAL_EVENTS = [
     {'type': 'ADDED', 'object': {'spec': 'a'}},
